fix: rename useApi -> useGetPositionApi
